My locations array doesn't have a fifth element, so locations[4] doesn't work.
Added a new object at the end of the locations array, following the same structure as the other objects. Set name to kill monster, set button text to an array with three Go to town square strings, set button functions to an array with three goTown variables, and set text to The monster screams Arg! as it dies. You gain experience points and find gold..