Note: i had to put a space before and after the efirst text quotation and on the second i had to put a space only before , not sure y.
Use the += operator to add Your <weapon> breaks., with a space in front of Your, to the end of text.innerText. Replace <weapon> with the last item in the inventory array using inventory.pop(), which will remove the last item in the array AND return it so it appears in your string.